# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API backend",
        "description": "Documentação da API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pagamentos": {
            "get": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Listar todos os pagamentos",
                "operationId": "aad1ee9bbaf6dfb2c9f6a3d552ce54e3",
                "responses": {
                    "200": {
                        "description": "Lista de pagamentos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pagamento"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Criar um novo pagamento",
                "operationId": "d5079fdccc372ef356fbbaaf74ea1ec4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PagamentoCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pagamento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pagamento"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pagamentos/{id}": {
            "get": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Mostrar detalhes de um pagamento",
                "operationId": "d0331b56874c68001c09f03e96f59a9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pagamento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagamento encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pagamento"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Pagamento não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Atualizar um pagamento existente",
                "operationId": "5b92aa9a30b7d72062091d5e5bfd3bbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pagamento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PagamentoUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pagamento atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pagamento"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Pagamento não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Excluir um pagamento",
                "operationId": "40d7cf9149d1bb33c3f56ceb511124f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pagamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagamento deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pagamento deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Pagamento não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pagamento": {
                "title": "Pagamento",
                "description": "Detalhes de um pagamento",
                "required": [
                    "id",
                    "valor",
                    "data_pagamento",
                    "metodo_pagamento",
                    "status",
                    "id_pedido"
                ],
                "properties": {
                    "id": {
                        "description": "ID único do pagamento",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "valor": {
                        "description": "Valor total do pagamento",
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "data_pagamento": {
                        "description": "Data em que o pagamento foi realizado",
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-23"
                    },
                    "metodo_pagamento": {
                        "description": "Método de pagamento (ex: cartão, boleto, pix)",
                        "type": "string",
                        "example": "cartao de credito"
                    },
                    "status": {
                        "description": "Status atual do pagamento (ex: pago, pendente, cancelado)",
                        "type": "string",
                        "example": "pago"
                    },
                    "id_pedido": {
                        "description": "ID do pedido ao qual este pagamento está associado",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Timestamp de criação do registro",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Timestamp da última atualização do registro",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "PagamentoCreateRequest": {
                "title": "PagamentoCreateRequest",
                "description": "Dados necessários para criar um novo pagamento",
                "required": [
                    "valor",
                    "data_pagamento",
                    "metodo_pagamento",
                    "status",
                    "id_pedido"
                ],
                "properties": {
                    "valor": {
                        "type": "number",
                        "format": "float",
                        "example": 250
                    },
                    "data_pagamento": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-24"
                    },
                    "metodo_pagamento": {
                        "type": "string",
                        "example": "pix"
                    },
                    "status": {
                        "type": "string",
                        "example": "pendente"
                    },
                    "id_pedido": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "PagamentoUpdateRequest": {
                "title": "PagamentoUpdateRequest",
                "description": "Dados opcionais para atualizar um pagamento existente",
                "properties": {
                    "valor": {
                        "type": "number",
                        "format": "float",
                        "example": 120.75
                    },
                    "data_pagamento": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-25"
                    },
                    "metodo_pagamento": {
                        "type": "string",
                        "example": "boleto"
                    },
                    "status": {
                        "type": "string",
                        "example": "cancelado"
                    },
                    "id_pedido": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": " * @OA\\SecurityScheme(\n * securityScheme=\"sanctum\",\n * type=\"http\",\n * scheme=\"bearer\",\n * bearerFormat=\"Sanctum\"\n * )\n ",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}